update:背包

+++ Assets/Scripts/Avatar/BaseClass/Resource.cs
@@ -39,6 +39,6 @@ public class Resource
 //资源集合(封装一下参数)
 public class resourceUnion
 {
-	public Dictionary<ResourceType,BigInteger> resourceConsumption;
+	public Dictionary<ResourceType,int> resourceConsumption;
 }
